requirements.txt
substack-api
beautifulsoup4
python-dotenv

SQLAlchemy>=2.0
psycopg2-binary

sentence-transformers
torch

groq

fastapi
uvicorn